Remove fakekube.
I was only using it one place anyway.
