+++ prow/cmd/sinker/main_test.go
@@ -17,12 +17,79 @@ limitations under the License.
 package main
 
 import (
+	"fmt"
 	"testing"
 	"time"
 
 	"k8s.io/test-infra/prow/kube"
-	"k8s.io/test-infra/prow/kube/fakekube"
 )
+
+type fakeClient struct {
+	Pods []kube.Pod
+	Jobs []kube.Job
+
+	DeletedPods []kube.Pod
+	DeletedJobs []kube.Job
+}
+
+func (c *fakeClient) ListPods(labels map[string]string) ([]kube.Pod, error) {
+	pl := make([]kube.Pod, 0, len(c.Pods))
+	for _, p := range c.Pods {
+		if labelsMatch(labels, p.Metadata.Labels) {
+			pl = append(pl, p)
+		}
+	}
+	return pl, nil
+}
+
+func (c *fakeClient) ListJobs(labels map[string]string) ([]kube.Job, error) {
+	jl := make([]kube.Job, 0, len(c.Jobs))
+	for _, j := range c.Jobs {
+		if labelsMatch(labels, j.Metadata.Labels) {
+			jl = append(jl, j)
+		}
+	}
+	return jl, nil
+}
+
+func (c *fakeClient) DeleteJob(name string) error {
+	for i, j := range c.Jobs {
+		if j.Metadata.Name == name {
+			c.Jobs = append(c.Jobs[:i], c.Jobs[i+1:]...)
+			c.DeletedJobs = append(c.DeletedJobs, j)
+			return nil
+		}
+	}
+	return fmt.Errorf("job %s not found", name)
+}
+
+
+func (c *fakeClient) DeletePod(name string) error {
+	for i, p := range c.Pods {
+		if p.Metadata.Name == name {
+			c.Pods = append(c.Pods[:i], c.Pods[i+1:]...)
+			c.DeletedPods = append(c.DeletedPods, p)
+			return nil
+		}
+	}
+	return fmt.Errorf("pod %s not found", name)
+}
+
+func labelsMatch(l1 map[string]string, l2 map[string]string) bool {
+	for k1, v1 := range l1 {
+		matched := false
+		for k2, v2 := range l2 {
+			if k1 == k2 && v1 == v2 {
+				matched = true
+				break
+			}
+		}
+		if !matched {
+			return false
+		}
+	}
+	return true
+}
 
 func TestClean(t *testing.T) {
 	pods := []kube.Pod{
@@ -167,7 +234,7 @@ func TestClean(t *testing.T) {
 		"old, deleted",
 		"old, aborted with pod",
 	}
-	kc := &fakekube.FakeClient{
+	kc := &fakeClient{
 		Pods: pods,
 		Jobs: jobs,
 	}
